+++ react/constants/SearchHelpers.js
@@ -156,6 +156,7 @@ export function getPagesArgs({
     default:
       break
   }
+  console.log("query", query)
 
   return {
     page: pagesPath,